hopefully the final collision refactor
