+++ src/lib/maths.ts
@@ -8,6 +8,7 @@ export * as MatrixExt from "./maths/matrixExt.js";
 export { default as Random } from "./maths/random.js";
 export * as RandomExt from "./maths/random.js";
 export { default as Rectangle } from "./maths/rectangle.js";
+export * as RectangleHelper from "./maths/rectangleHelper.js";
 export { default as Vector2 } from "./maths/vector2.js";
 export * as Vector2Ext from "./maths/vector2Ext.js";
 export { default as Vector3 } from "./maths/vector3.js";
